requirements.txt
PyQt6
speechmatics-python
requests
tqdm